Cambio de template y css añadidos a vistas de login de usuarios y divisiones, corregido error de logica al eliminar usuarios y refactorizacion de errores y de layouts, añadida imagen de background.

+++ app/Http/Controllers/UserInfoController.php
@@ -24,8 +24,11 @@ class UserInfoController extends Controller
         foreach ($user->number as $number) {
             $number->delete();
         }
+
         $port = Portability::where('user_id', $user->id)->first();
-        $port->delete();
+        if ($port != null) {
+            $port->delete();
+        }
         $user->delete();
 
         session()->flash('message', 'User deleted succesfuly');
